+++ python/rmm/_lib/device_buffer.pyx
@@ -13,11 +13,11 @@ cdef class DeviceBuffer:
                   uintptr_t ptr=0,
                   size_t size=0,
                   uintptr_t stream=0):
-        cdef void* c_ptr
+        cdef const void* c_ptr
         cdef cudaStream_t c_stream
 
         with nogil:
-            c_ptr = <void*>ptr
+            c_ptr = <const void*>ptr
             c_stream = <cudaStream_t>stream
 
             if c_ptr == NULL:
@@ -26,6 +26,9 @@ cdef class DeviceBuffer:
                 self.c_obj.reset(new device_buffer(c_ptr, size, c_stream))
 
     def __len__(self):
+        return self.size
+
+    def __sizeof__(self):
         return self.size
 
     @property