+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 
 ## New Features
 
+- PR #252 Add `__sizeof__` method to `DeviceBuffer`
+
 ## Improvements
 
 - PR #246 Type `DeviceBuffer` arguments to `__cinit__`
 - PR #249 Use `DeviceBuffer` in `device_array`
 - PR #256 Cast through `uintptr_t` to `cudaStream_t`
+- PR #254 Use `const void*` in `DeviceBuffer.__cinit__`
 
 ## Bug Fixes